feat: add unit test for sqlite db
