+++ models/repo/message_repo.go
@@ -28,6 +28,7 @@ type MessageRepo interface {
 	GetSignedMessageByTime(start time.Time) ([]*types.Message, error)
 	GetSignedMessageByHeight(height abi.ChainEpoch) ([]*types.Message, error)
 	GetSignedMessageFromFailedMsg(addr address.Address) ([]*types.Message, error)
+
 	ListMessage() ([]*types.Message, error)
 	ListMessageByFromState(from address.Address, state types.MessageState, isAsc bool, pageIndex, pageSize int) ([]*types.Message, error)
 	ListMessageByAddress(addr address.Address) ([]*types.Message, error)
